refactor: 워크스페이스 메인 유저 매핑 컴포넌트로 변경
